sedcli: enhancements in get_password
This patch introduces following changes:
 * min and max password length requirements can be passed to
get_password() function. It will reject invalid input
 * sedcli from now on accepts password min 8 characters in length
 * get_password() has been moved to separate compilation unit allowing
its reuse in different binaries
