+++ src/sedcli_util.h
@@ -1,0 +1,6 @@
+#ifndef _SEDCLI_UTIL_H_
+#define _SEDCLI_UTIL_H_
+
+int get_password(char *pwd, uint8_t *len, uint8_t min, uint8_t max);
+
+#endif /* _SEDCLI_UTIL_H_ */
